+++ src/pages/home/sections/MicroTasksQueue/MicroTasksQueue.tsx
@@ -1,7 +1,7 @@
 import { useQueueManagerStore } from 'store/store.ts';
 import * as Styled from './MicroTasksQueue.styled.ts';
 import InfoIcon from 'components/InfoIcon/InfoIcon.tsx';
-import useBoolean from 'utils/useBoolean.tsx';
+import useBoolean from '../../../../hooks/useBoolean.tsx';
 import { Zoom } from '@mui/material';
 import { List } from '../../Home.styled.ts';
 import MicroTasksQueueModal from './MicroTasksQueue.modal.tsx';
