+++ src/utils/constants.ts
@@ -1,3 +1,1 @@
-export const MIN_DESKTOP_WIDTH = 768;
-
 export const EVENT_LOOP_ID = 'eventLoopId';
