--- src/utils/isMobile.ts
@@ -1,3 +1,0 @@
-import { MIN_DESKTOP_WIDTH } from './constants';
-
-export const isMobile = () => window.innerWidth < MIN_DESKTOP_WIDTH;
